Functional game
You are able to move the ship side to side, when the space bar is pressed you shoot a laser, and asteroids falling from the sky are auto-generated in the script "asteroidsSpawner". the laser destroys itself when out of the screen or when in contact with an asteroid; the asteroid destroys itself when in contact with a laser; and the ship destroys itself when in contact with an asteroid. 

next: needs to improver spawner time, understand it better, and linearly increase difficult (the more asteroids you break, the faster it respawns); add a score keeper; add a game over/play again when the ship is hit.