score UI ,increasing difficulty
Added a score UI that is incremented every time a player destroys an asteroid in ScoreScript.cs. refers to the score's text box, and has a public method incrementScore(), gets the current value in the text box, and adds 1; incrementScore() is called in AsteroidCollider.OnCollisionEnter().

In AsteroidSpawner: 
adjusted spawn() so that -timer is increased by 1 every second. 
implemented static increaseDifficulty(), decreases -maxDelay and -minDelay, making asteroids spawn faster, increasing difficulty. the method is called in AsteroidCollider.OnCollisionEnter()